BO para inscripción con pago en efectivo OK

+++ ProyectoJavaCERO/CERO/Persistencia/src/main/java/Entidades/MetodoPago.java
@@ -5,17 +5,17 @@ package Entidades;
  * @author Usuario
  */
 public abstract class MetodoPago {
-    private Long id;
+    private Integer id;
 
-    public MetodoPago(Long id) {
+    public MetodoPago(Integer id) {
         this.id = id;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
